feat(database): 统一金额字段为分单位并重构商品订单表结构
- 修改 mall-elk.conf 文件增加空行以优化格式
- 新增 amount_unify_to_fen.sql 脚本实现全链路金额统一为分（bigint）
- 新增 p0_product_sku_order_cart_refactor.sql 脚本重构商品SKU订单购物车表结构
- 更新 README.md 文档增加金额统一说明和相关SQL链接
- 在商品SKU设计文档中增加落地记录和金额单位说明
- 修改 SysUser 相关实体类将 balance 字段从 BigDecimal 改为 Long 类型
- 更新 SysUserService 接口和实现类中的金额参数类型为 Long
- 调整用户余额扣减和充值方法的金额比较逻辑
- 更新测试代码中的余额初始化和断言逻辑
- 在网关过滤器中优化 SkyWalking 链路追踪的 MDC 设置逻辑
- 配置 logback-spring.xml 将链路追踪信息输出到日志和 ELK

+++ mall-admin/mall-sys-core/src/main/java/com/yz/mall/sys/service/SysUserService.java
@@ -12,7 +12,6 @@ import com.yz.mall.sys.vo.SysTreeMenuVo;
 import com.yz.mall.sys.vo.SysUserVo;
 import jakarta.validation.Valid;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -91,16 +90,16 @@ public interface SysUserService extends IService<SysUser> {
      * 扣减账户金额
      *
      * @param userId 扣减用户
-     * @param amount 扣减金额
+     * @param amount 扣减金额（单位：分）
      */
-    void deduct(Long userId, BigDecimal amount);
+    void deduct(Long userId, Long amount);
 
     /**
      * 账户充值
      *
      * @param userId 充值用户
-     * @param amount 充值金额
+     * @param amount 充值金额（单位：分）
      */
-    void recharge(Long userId, BigDecimal amount);
+    void recharge(Long userId, Long amount);
 }
