+++ mall-admin/mall-sys-feign/src/main/java/com/yz/mall/sys/service/impl/ExtendSysUserServiceImpl.java
@@ -13,7 +13,6 @@ import com.yz.mall.sys.vo.ExtendSysUserSlimVo;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -33,7 +32,7 @@ public class ExtendSysUserServiceImpl implements ExtendSysUserService {
     }
 
     @Override
-    public void deduct(Long userId, BigDecimal amount) {
+    public void deduct(Long userId, Long amount) {
         Result<Boolean> result = feign.deduct(new ExtendSysUserBalanceDto(userId, amount));
         if (!CodeEnum.SUCCESS.get().equals(result.getCode())) {
             throw new FeignException(result.getCode(), result.getMsg());
@@ -41,7 +40,7 @@ public class ExtendSysUserServiceImpl implements ExtendSysUserService {
     }
 
     @Override
-    public void recharge(Long userId, BigDecimal amount) {
+    public void recharge(Long userId, Long amount) {
         Result<Boolean> result = feign.deduct(new ExtendSysUserBalanceDto(userId, amount));
         if (!CodeEnum.SUCCESS.get().equals(result.getCode())) {
             throw new FeignException(result.getCode(), result.getMsg());
