+++ mall-admin/mall-sys-interface/src/main/java/com/yz/mall/sys/service/ExtendSysUserService.java
@@ -7,7 +7,6 @@ import com.yz.mall.sys.vo.ExtendSysUserSlimVo;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -23,17 +22,17 @@ public interface ExtendSysUserService {
      * 扣减账户金额
      *
      * @param userId 扣减用户
-     * @param amount 扣减金额
+     * @param amount 扣减金额（单位：分）
      */
-    void deduct(Long userId, BigDecimal amount);
+    void deduct(Long userId, Long amount);
 
     /**
      * 账户充值
      *
      * @param userId 充值用户
-     * @param amount 充值金额
+     * @param amount 充值金额（单位：分）
      */
-    void recharge(Long userId, BigDecimal amount);
+    void recharge(Long userId, Long amount);
 
     /**
      * 获取指定用户所拥有的角色
