+++ mall-pms/mall-pms-dao/src/main/java/com/yz/mall/pms/vo/PmsProductDisplayInfoVo.java
@@ -5,7 +5,6 @@ import com.yz.mall.pms.entity.PmsStock;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -30,9 +29,9 @@ public class PmsProductDisplayInfoVo implements Serializable {
     private String productName;
 
     /**
-     * 商品价格
+     * 商品价格（单位：分）
      */
-    private BigDecimal productPrice;
+    private Long productPrice;
 
     /**
      * 商品标签
