+++ mall-pms/mall-pms-interface/src/main/java/com/yz/mall/pms/dto/ExtendPmsCartDto.java
@@ -3,7 +3,6 @@ package com.yz.mall.pms.dto;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 /**
  * 购物车指定商品的Id和数量
@@ -44,13 +43,13 @@ public class ExtendPmsCartDto implements Serializable {
     private Integer quantity;
 
     /**
-     * 商品优惠金额
+     * 商品优惠金额（单位：分）
      */
-    private BigDecimal discountAmount;
+    private Long discountAmount;
 
     /**
-     * 商品优惠后的实际价格
+     * 商品优惠后的实际价格（单位：分）
      */
-    private BigDecimal realAmount;
+    private Long realAmount;
 }
